Rename backend/utils.py to db.py
