+++ dtbase/backend/api/location/routes.py
@@ -10,6 +10,7 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from dtbase.backend.auth import authenticate_access
+from dtbase.backend.db import db_session
 from dtbase.backend.models import (
     Coordinates,
     LocationIdentifier,
@@ -17,7 +18,6 @@ from dtbase.backend.models import (
     MessageResponse,
     ValueType,
 )
-from dtbase.backend.utils import db_session
 from dtbase.core import locations
 from dtbase.core.exc import RowExistsError, RowMissingError
