+++ dtbase/backend/api/model/routes.py
@@ -10,6 +10,7 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from dtbase.backend.auth import authenticate_access
+from dtbase.backend.db import db_session
 from dtbase.backend.models import (
     MessageResponse,
     Model,
@@ -18,7 +19,6 @@ from dtbase.backend.models import (
     SensorMeasureIdentifier,
     ValueType,
 )
-from dtbase.backend.utils import db_session
 from dtbase.core import models
 from dtbase.core.exc import RowMissingError
