+++ dtbase/backend/api/sensor/routes.py
@@ -10,8 +10,8 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from dtbase.backend.auth import authenticate_access
+from dtbase.backend.db import db_session
 from dtbase.backend.models import MessageResponse, SensorMeasure, SensorType, ValueType
-from dtbase.backend.utils import db_session
 from dtbase.core import sensor_locations, sensors
 from dtbase.core.exc import RowMissingError
