+++ dtbase/backend/auth.py
@@ -9,8 +9,8 @@ from fastapi.security import OAuth2PasswordBearer
 from pydantic import ValidationError
 from sqlalchemy.orm import Session
 
+from dtbase.backend.db import db_session
 from dtbase.backend.models import ParsedToken, TokenPair
-from dtbase.backend.utils import db_session
 from dtbase.core import users
 from dtbase.core.constants import (
     JWT_ACCESS_TOKEN_EXPIRES,
