+++ dtbase/core/locations.py
@@ -3,8 +3,8 @@ from typing import Any, Dict, List, Optional, cast
 
 import sqlalchemy as sqla
 
-from dtbase.backend.utils import Session
 from dtbase.core import queries, utils
+from dtbase.core.db import Session
 from dtbase.core.exc import RowExistsError, RowMissingError, TooManyRowsError
 from dtbase.core.structure import (
     Location,
